+++ earthdaily/earthdatastore/cube_utils/__init__.py
@@ -263,6 +263,12 @@ def _disable_known_datacube_warning():
             message="invalid value encountered in cast",
             module="dask.array.chunk",
         )
+        warnings.filterwarnings(
+            "ignore", category=RuntimeWarning, message="All-NaN slice encountered"
+        )
+        warnings.filterwarnings(
+            "ignore", category=RuntimeWarning, message="Mean of empty slice"
+        )
 
 
 def datacube(
@@ -385,7 +391,7 @@ def datacube(
                 json.dumps(ds.coords[coord].values[idx])
                 for idx in range(ds.coords[coord].size)
             ]
-
+    ds = ds.sortby(ds.time)
     return ds
 
 
@@ -503,7 +509,7 @@ def rescale_assets_with_items(
             # Track scaling parameters
             scales.setdefault(ds_asset, {}).setdefault(scale, {}).setdefault(
                 offset, []
-            ).append(time)
+            ).append(idx)
 
     # Apply rescaling
     if scales:
@@ -512,8 +518,7 @@ def rescale_assets_with_items(
             asset_scaled = []
             for scale, offset_data in scale_data.items():
                 for offset, times in offset_data.items():
-                    mask = np.isin(ds.time, times)
-                    asset_scaled.append(ds[[asset]].isel(time=mask) * scale + offset)
+                    asset_scaled.append(ds[[asset]].isel(time=times) * scale + offset)
             scaled_assets.append(xr.concat(asset_scaled, dim="time"))
 
         # Merge scaled assets