+++ docs/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2025-02-13
+
+### Added
+
+- New default method for duplicates detection using `proj:transform` instead
+of bbox (huge gap). Change can be done using 
+`filter_duplicate_items(...,method='bbox')`.
+
+### Fixed
+
+- Issues with rescale when same datetime between products.
+
 ## [0.5.1] - 2025-02-12
 
 ### Added